Prevent "region not set" errors
This probably isn't the right fix, but it resolves an issue that folks run into
where the aws provider in the modules doesn't know what region to use.

I think the real issue is that we need to use the provider defined in main.tf
and "pass" it to the modules.