Explicitly pass the provider to the modules
In order to prevent "Provider configuration not present" error. I think this is
because way back, the modules specified their own provider which is not
recommended by Terraform (use the `required_providers` instead).

However, simply removing them doesn't seem to be enough. It seems we'll run into
this whenever we destroy an old resource. Being explicit here prevents that.